ipfs_api/requirements.txt
CairoSVG
fastapi
pinata
python-dotenv
uvicorn
